:label: [rename] domain/CategoryType을 response/HomeResponse 내부로 이동 및 오전/오후 시간 표기 변경
